service_backend/requirements/requirements.txt
# -- linting
black==21.9b0 # devdepenency

# -- postgresql
psycopg2==2.8.6

# -- django
asgiref==3.3.4
Django==3.2.9
pytz==2021.1
sqlparse==0.4.1

# -- third party
django-phonenumber-field==5.0.0
phonenumberslite==8.12.21
Faker==10.0.0 # devdepenency

# -- django rest framework
djangorestframework==3.12.4
Markdown==3.3.4
django-cors-headers==3.7.0

# -- djoser
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
uritemplate==3.0.1
urllib3==1.26.4




